requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
tavily-python>=0.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
